Loading screen renamed to Splash; Placed logOut function from useAuthUpdate in Explore; ResetPassword takes resetPassword function from useAuthUpdate to generate a custom Alert that takes user back to login screen onPress
