+++ app/screens/ExploreTab.js
@@ -1,10 +1,14 @@
 import React from 'react';
-import { StyleSheet, View, Text } from 'react-native';
+import { StyleSheet, View, Text, Button } from 'react-native';
+import { useAuthUpdate } from '../hooks/useAuth'
 
 function ExploreTab() {
+  const { logOut } = useAuthUpdate();
+
   return (
     <View style={styles.container}>
       <Text style={styles.text}>Explore Tab</Text>
+      <Button title="Log Out" onPress={() => logOut()} />
     </View>
   );
 }
